requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
boto3==1.37.10; python_version >= '3.8'
botocore==1.37.10; python_version >= '3.8'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
coreapi==2.3.3
coreschema==0.0.4
django==5.1.7; python_version >= '3.10'
django-debug-toolbar==5.0.1; python_version >= '3.9'
django-rest-swagger==2.2.0
django-semantic-admin==0.6.6; python_version >= '3.10'
django-semantic-forms==0.1.8; python_version >= '3.10'
django-storages==1.14.5; python_version >= '3.7'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.5.0; python_version >= '3.9'
drf-yasg==1.21.10; python_version >= '3.6'
idna==3.10; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.6; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
openapi-codec==1.3.2
packaging==24.2; python_version >= '3.8'
pillow==11.1.0; python_version >= '3.9'
pyjwt==2.9.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2025.1
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
s3transfer==0.11.4; python_version >= '3.8'
simplejson==3.20.1; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.3; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.3.0; python_version >= '3.10'
